Add the ability to build multiple examples (#15)
This adds the ability to have multiple examples in a package. If there is no pubspec.yaml in the example directory, then it looks down one more level to find any pubspecs.

Also added a message if the user forgets to specify one of --apk or --ipa when building examples.