Support `integration_test/` directory for `drive-examples` command
For any driver script in `packages/$package/example/test_driver` that ends with `_test.dart`, this plugin will attempt to use this driver script to drive all the files in `packages/$package/example/integration_test/` that end with `_test.dart`.

This also changes the drive CLI command to pass `--target` and `--driver` explicitly, which is a no-op change because this is how the files are inferred by the flutter tool.